+++ packages/cryptoplease/lib/app/screens/authenticated/profile/components/profile_section.dart
@@ -10,7 +10,7 @@ import '../../../../../di.dart';
 import '../../../../../features/incoming_split_key_payments/module.dart';
 import '../../../../../features/legal/flow.dart';
 import '../../../../../features/outgoing_direct_payments/module.dart';
-import '../../../../../features/outgoing_split_key_payments/bl/repository.dart';
+import '../../../../../features/outgoing_split_key_payments/module.dart';
 import '../../../../../features/payment_request/bl/repository.dart';
 import '../../../../../gen/assets.gen.dart';
 import '../../../../../l10n/l10n.dart';
