+++ packages/cryptoplease/lib/data/db/db.dart
@@ -3,7 +3,7 @@ import 'package:injectable/injectable.dart';
 
 import '../../features/incoming_split_key_payments/module.dart';
 import '../../features/outgoing_direct_payments/module.dart';
-import '../../features/outgoing_split_key_payments/bl/repository.dart';
+import '../../features/outgoing_split_key_payments/module.dart';
 import '../../features/payment_request/bl/repository.dart';
 import 'open_connection.dart';
