+++ packages/cryptoplease/lib/features/outgoing_split_key_payments/module.dart
@@ -2,10 +2,16 @@ import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:nested/nested.dart';
 import 'package:provider/provider.dart';
+import 'package:uuid/uuid.dart';
 
 import '../../core/accounts/bl/account.dart';
+import '../../core/amount.dart';
 import '../../di.dart';
-import 'bl/bloc.dart';
+import 'src/bl/bloc.dart';
+
+export 'src/bl/oskp_verifier.dart';
+export 'src/bl/outgoing_split_key_payment.dart';
+export 'src/bl/repository.dart';
 
 class OSKPModule extends SingleChildStatelessWidget {
   const OSKPModule({Key? key, Widget? child}) : super(key: key, child: child);
@@ -21,4 +27,14 @@ class OSKPModule extends SingleChildStatelessWidget {
         ],
         child: child,
       );
+}
+
+extension BuildContextExt on BuildContext {
+  String createOSKP({required CryptoAmount amount}) {
+    final id = const Uuid().v4();
+    final event = OSKPEvent.create(amount: amount, id: id);
+    read<OSKPBloc>().add(event);
+
+    return id;
+  }
 }
