+++ packages/cryptoplease/lib/features/outgoing_split_key_payments/src/bl/link_shortener.dart
