+++ packages/cryptoplease/lib/features/outgoing_split_key_payments/src/bl/outgoing_split_key_payment.dart
@@ -3,7 +3,7 @@ import 'package:freezed_annotation/freezed_annotation.dart';
 import 'package:solana/encoder.dart';
 import 'package:solana/solana.dart';
 
-import '../../../core/amount.dart';
+import '../../../../core/amount.dart';
 
 part 'outgoing_split_key_payment.freezed.dart';
