+++ packages/cryptoplease/lib/features/outgoing_split_key_payments/src/presentation/oskp_confirmation_screen.dart
@@ -2,18 +2,18 @@ import 'package:auto_route/auto_route.dart';
 import 'package:dfunc/dfunc.dart';
 import 'package:flutter/material.dart';
 
-import '../../../core/amount.dart';
-import '../../../core/conversion_rates/presentation/conversion_rates.dart';
-import '../../../core/currency.dart';
-import '../../../core/presentation/format_amount.dart';
-import '../../../l10n/device_locale.dart';
-import '../../../l10n/l10n.dart';
-import '../../../ui/app_bar.dart';
-import '../../../ui/button.dart';
-import '../../../ui/chip.dart';
-import '../../../ui/content_padding.dart';
-import '../../../ui/info_widget.dart';
-import '../../../ui/theme.dart';
+import '../../../../core/amount.dart';
+import '../../../../core/conversion_rates/presentation/conversion_rates.dart';
+import '../../../../core/currency.dart';
+import '../../../../core/presentation/format_amount.dart';
+import '../../../../l10n/device_locale.dart';
+import '../../../../l10n/l10n.dart';
+import '../../../../ui/app_bar.dart';
+import '../../../../ui/button.dart';
+import '../../../../ui/chip.dart';
+import '../../../../ui/content_padding.dart';
+import '../../../../ui/info_widget.dart';
+import '../../../../ui/theme.dart';
 
 class OSKPConfirmationScreen extends StatelessWidget {
   const OSKPConfirmationScreen({
