+++ packages/cryptoplease/lib/features/pending_activities/pending_activity.dart
@@ -1,7 +1,7 @@
 import 'package:freezed_annotation/freezed_annotation.dart';
 
 import '../outgoing_direct_payments/module.dart';
-import '../outgoing_split_key_payments/bl/outgoing_split_key_payment.dart';
+import '../outgoing_split_key_payments/module.dart';
 
 part 'pending_activity.freezed.dart';
